Address Gemini PR review comments
GEMINI FEEDBACK IMPLEMENTATION:

1. Documentation Clarity (docs/dev/cli-restructure.md):
   - Added note to SERVICE Group section clarifying that only 'install'
     subcommand is currently implemented
   - Other service management commands (start/stop/status/restart/logs)
     are planned for future releases
   - Improves clarity for future developers about current implementation state

2. Avoid Circular Dependencies (CLI command registry):
   - Created separate src/mmrelay/cli_utils.py module for command registry
   - Moved CLI_COMMANDS dict and all utility functions from cli.py
   - Updated imports in cli.py, config.py, and matrix_utils.py
   - Both cli.py and config.py can now import from cli_utils without circular deps
   - Maintains single source of truth for all CLI command references

BENEFITS:
- Better code organization with clear separation of concerns
- Eliminates potential circular import issues
- Maintains all existing functionality while improving architecture
- Clear documentation about current vs planned CLI features
